requirements.txt
Django==5.0.3
django-environ==0.11.2
argon2-cffi-bindings==21.2.0
psycopg2==2.9.9
argon2-cffi==23.1.0



django-cors-headers==4.3.1
djangorestframework==3.15.1
